+++ hyperparameter_hunter/__init__.py
@@ -8,9 +8,19 @@ try:
     hook_keras_layer()
 except DistributionNotFound:
     pass
-except Exception as _ex:
+except Exception:
     raise
 
+##################################################
+# Store Library Version
+##################################################
+import os.path
+
+try:
+    with open(os.path.join(os.path.dirname(__file__), "VERSION")) as f:
+        __version__ = f.read().strip()
+except Exception:
+    raise
 
 ##################################################
 # Set __all__
@@ -18,9 +28,9 @@ except Exception as _ex:
 from .environment import Environment
 from .experiments import CrossValidationExperiment
 from .optimization import BayesianOptimization
-from .optimization import GradientBoostedRegressionTreeOptimization
-from .optimization import RandomForestOptimization
-from .optimization import ExtraTreesOptimization
+from .optimization import GradientBoostedRegressionTreeOptimization, GBRT
+from .optimization import RandomForestOptimization, RF
+from .optimization import ExtraTreesOptimization, ET
 from .optimization import DummySearch
 from .space import Real
 from .space import Integer
@@ -35,8 +45,11 @@ __all__ = [
     #################### Hyperparameter Optimization ####################
     "BayesianOptimization",
     "GradientBoostedRegressionTreeOptimization",
+    "GBRT",
     "RandomForestOptimization",
+    "RF",
     "ExtraTreesOptimization",
+    "ET",
     "DummySearch",
     #################### Search Space ####################
     "Real",